% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bray.curtis.activity.R
\name{bray.curtis.activity}
\alias{bray.curtis.activity}
\title{Gets Bray Curtis activity for a popdy data object}
\usage{
bray.curtis.activity(x, scale = TRUE)
}
\arguments{
\item{x}{a popdy data object}

\item{scale}{flag to specify whether the output is to be scaled by the population}
}
\description{
Gets Bray Curtis activity for a popdy data object
}
\keyword{evolution}
\keyword{qnn}

