% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecea.R
\name{ecea.activity}
\alias{ecea.activity}
\title{Gets Ed Clark's activity for a popdy data object}
\usage{
ecea.activity(x, scale = FALSE)
}
\arguments{
\item{x}{a popdy data object}

\item{scale}{flag to specify whether the output is to be scaled by the population}
}
\description{
Gets Ed Clark's activity for a popdy data object
}
\keyword{evolution}
\keyword{qnn}

