% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{get_qnns}
\alias{get_qnns}
\title{Gets QNN summary data for a set of files and stores the data in an output file}
\usage{
get_qnns(stem = "popdy", N = 1, outstem = "qnn")
}
\arguments{
\item{stem}{a string for the root name of the file}

\item{N}{the number of files to be processed}

\item{outstem}{a string for the root of the output name of the file}
}
\description{
This function allows you to calculate QNN for a set of files with the same name
for example, if your files are called 'pop001.dat', 'pop002.dat'
}
\examples{
get_qnns("pop",3,"qnn")
reads files 'pop1.dat', 'pop2.dat' and 'pop3.dat',
and outputs files 'qnn1.dat', 'qnn2.dat' and 'qnn3.dat'
}
\keyword{evolution}
\keyword{qnn}

