% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{nspecies}
\alias{nspecies}
\title{get the number of species in a popdy object}
\usage{
nspecies(x)
}
\arguments{
\item{x}{a popdy data object}
}
\description{
get the number of species in a popdy object
}
\keyword{evolution}
\keyword{qnn}

