% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdy.totextfile.R
\name{popdy.totextfile}
\alias{popdy.totextfile}
\title{Exports a popdy object to a text file.
Each line in the text file is of format 't,s,p' where t=time,s=species,p=population}
\usage{
popdy.totextfile(x = NA, fn = "popdy.dat")
}
\arguments{
\item{x}{a popdy data object}

\item{fn}{the output file name (default is 'popdy.dat'}
}
\description{
Exports a popdy object to a text file.
Each line in the text file is of format 't,s,p' where t=time,s=species,p=population
}
\keyword{evolution}
\keyword{qnn}

