% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotpopdy.R
\name{popdyplot}
\alias{popdyplot}
\title{plots a  popdy data object}
\usage{
popdyplot(x, xlab = "Time", ylab = "Count", plot.zero = FALSE,
  col.fun = rainbow)
}
\arguments{
\item{x}{a popdy data object}

\item{xlab}{The label on the x-axis (default = "Time")}

\item{ylab}{The label on the y-axis (default = "Count")}

\item{plot.zero}{flag to plot zero-valued columns (default = FALSE) )}

\item{col.fun}{specifies the method for generating the line colours (default = "rainbow") )}
}
\description{
plots a  popdy data object
}
\keyword{evolution}
\keyword{qnn}

