% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{qnn.activity}
\alias{qnn.activity}
\title{Gets QNN activity for a popdy data object}
\usage{
qnn.activity(x, scale = FALSE)
}
\arguments{
\item{x}{a popdy data object}

\item{scale}{flag to specify whether the output is to be scaled by the population}
}
\description{
Gets QNN activity for a popdy data object
}
\keyword{evolution}
\keyword{qnn}

