% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{qnn.summarize}
\alias{qnn.summarize}
\title{summarises QNN activity into a single figure}
\usage{
qnn.summarize(x)
}
\arguments{
\item{x}{a popdy data object holding the qnn values (rather than the popdy values)}
}
\description{
summarises QNN activity into a single figure
}
\keyword{evolution}
\keyword{qnn}

