% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randbow.R
\name{randbow}
\alias{randbow}
\title{creates an array of colors, randomly permuted from a rainbow}
\usage{
randbow(x)
}
\arguments{
\item{x}{a popdy data object}
}
\description{
creates an array of colors, randomly permuted from a rainbow
}
\keyword{color}
\keyword{colour}

