% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{times}
\alias{times}
\title{get a vector of times from a popdy object}
\usage{
times(x)
}
\arguments{
\item{x}{a popdy data object}
}
\description{
get a vector of times from a popdy object
}
\keyword{evolution}
\keyword{qnn}

