% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genqnn.R
\name{read.popdy}
\alias{read.popdy}
\title{read a file into a popdy object}
\usage{
read.popdy(file)
}
\arguments{
\item{file}{a comma-delimited file, each line of which details time,species,count}
}
\description{
read a file into a popdy object
}
\keyword{evolution}
\keyword{qnn}

